Cleanup

+++ src/genie/format/mgb/exporter.cc
@@ -28,7 +28,7 @@ void Exporter::flowIn(core::AccessUnit&& t, const util::Section& id) {
     core::AccessUnit data = std::move(t);
     util::OrderedSection section(&lock, id);
     getStats().add(data.getStats());
-    size_t parameter_id = parameter_stash.size();
+    uint8_t parameter_id = parameter_stash.size();
     core::parameter::ParameterSet out_set(parameter_id, parameter_id, std::move(data.getParameters()));
     mgb::RawReference ref;
     for (const auto& p : data.getRefToWrite()) {
@@ -70,8 +70,8 @@ void Exporter::flowIn(core::AccessUnit&& t, const util::Section& id) {
                            : (data.isReferenceOnly() ? core::parameter::DataUnit::DatasetType::REFERENCE
                                                      : core::parameter::DataUnit::DatasetType::NON_ALIGNED);
 
-    mgb::AccessUnit au((uint32_t)id_ctr, (uint8_t)parameter_id, data.getClassType(), (uint32_t)data.getNumReads(),
-                       datasetType, 32, false, core::AlphabetID::ACGTN);
+    mgb::AccessUnit au((uint32_t)id_ctr, parameter_id, data.getClassType(), (uint32_t)data.getNumReads(), datasetType,
+                       32, false, core::AlphabetID::ACGTN);
     if (data.isReferenceOnly()) {
         au.getHeader().setRefCfg(RefCfg(data.getReference(), data.getReferenceExcerpt().getGlobalStart(),
                                         data.getReferenceExcerpt().getGlobalEnd() - 1, 32));
